Handling for module-level dimension variables for interstitial variables (#683)
Updates to allow for dimension variables that are set in the register
phase to be used to allocate interstitial variables.

Open to feedback, of course! There's no explicit handling right now to
_ensure_ that the variable is set in the register phase, but my feeling
is that you'll get a run-time error if you try to use an uninitialized
variable in an `allocate` statement.

User interface changes?: No

Fixes: 
closes #681

Testing: Added variables/logic in capgen test to confirm the dimensions
are initialized and used properly

---------

Co-authored-by: Dom Heinzeller <dom.heinzeller@icloud.com>