Improve parsing of Fortran array specifications (#687)
Improve parsing of Fortran array specifications

capgen was reporting odd errors when a semi-colon typo appeared in the
middle of an
array specification.
This is due to the fact that the semi-colon is a Fortran special
character used to separate statements making it possible to have
multiple statements on a single source line.
This PR allows semi-colons to appear in parentheses in a statement.
To have this result in a more useful parsing error, some refactoring was
done on how the Fortran parser collects and displays errors (an onion
peel I have wanted to do for a while although there is plenty of onion
left).

User interface changes?: No

Fixes: #362 

Testing:
  test removed:
  unit tests:
  system tests:
  manual testing: