Add capability to parse nested suites in capgen (#691)
This PR adds the capability to parse nested suites in capgen, as
discussed in https://github.com/NCAR/ccpp-framework/issues/275.

**Notes/Features**
1. The nested suites must reside in separate files, because only one
`<suite>` root element can exist in each file.
2. Suites can be nested recursively.
3. Old suite definition files using XML schema 1.0 are still supported,
but nested suites are only support for schema version 2 and above.
4. Nested suites inside a group construct **must** have the `group=`
attribute; the content of that group in the referenced suite is merged
into the existing group while preserving the order. This is the first
example (aka the merge=true example) described in
https://github.com/NCAR/ccpp-framework/issues/275#issue-589991333.
5. Nested suites one level up, i.e. at the group level, may or may not
have the `group=` attribute; if they do, only this group is imported,
otherwise all groups are imported in the correct location in the main
suite. This is the second example (aka the merge=false example)
described in
https://github.com/NCAR/ccpp-framework/issues/275#issue-589991333.
6. Nested suites can be recursive, with each nested suite in its
separate file.
7. There is currently a hardcoded limit of 10 nested suites that can be
referenced in a suite at either of the two levels described in 4 and 5,
including recursive nested suites. This is to allow catching infinite
recursions.
8. **Disclaimer** I don't like writing docstrings and doctests,
therefore I asked chatgpt to do that for me. It did a decent job,
certainly better than I would have, but I still had to go in and correct
a few things.
9. The `PrettyElementTree` class was replaced by a much shorter function
`write_xml_file` that uses features available in the already-used XML
python library. And since it contained `dom` twice in the name, I simply
couldn't resist! The output of the new implementation is almost
identical to the previous `PrettyElementTree` output, see screenshot at
the bottom of the PR description.
10. More than 90% of the changed/added files in this PR are for the new
test `nested_suite_test` and the new unit test suite `test_sdf.py`, thus
the actual changes to the code are quite small and hopefully easy to
review.

User interface changes?: **Yes, but these are optional.** In order to
use the new functionality, users have to update their suite definition
file to the new XML schema version 2.0 and use the new syntax for
`nested_suite` elements. There are no user interface changes for the
previous XML schema 1.0, which remains valid, and there are no user
interface changes for invoking capgen or in the auto-generated code.

## Issues
Fixes #275

## Testing

Test removed: none
Test added:
- added `test/nested_suite_test` - see `README.md` in that directory for
more information
    - added doctests for the new functions in `xml_tools.py`
- added unit tests in `test/unit_test/test_sdf.py` (credit @gold2718)
 Unit tests: all pass
 System tests: all pass
 Manual testing: `nested_suite_test`, `test_sdf.py`, new docstring tests

## Additional information

Difference in `datatable.xml` output of the new `write_xml_file`
function (left) and the old `PrettyElementTree` class (right):

<img width="2528" height="1668" alt="image"
src="https://github.com/user-attachments/assets/0db744ef-2c3d-4ccc-9c52-2ecf5e9e973e"
/>

In a nutshell, the differences are no white spaces before the closing
characters of XML elements, and the additional `<? xml ...?>` line at
the top. Both of these seem reasonable to me.

---------

Co-authored-by: Steve Goldhaber <stevenng@met.no>